software/python/citaComms/requirements.txt
pyserial>=3.0.1
mongo==3.0.3
